Cleaned up old code
